model/teacher/ngp_pl/requirements.txt
einops==0.4.1
kornia==0.6.5
pytorch-lightning==1.7.7
matplotlib==3.5.2
opencv-python==4.6.0.66
lpips
imageio==2.24
imageio-ffmpeg
jupyter
scipy
pymcubes
trimesh
dearpyguipip 
torchmetrics==0.11.4